refs #24 minor modifications to code to get rid of some nuggets. Initial commit to initiate proper subversion on C files -- again
git-svn-id: https://svn.isis.rl.ac.uk/Horace/trunk@527 1cb955da-4463-431d-88dd-2af2f5550de3

+++ _LowLevelCode/cpp/sort_pixels_by_bins/sort_pixels_by_bins/sort_pixels_by_bins.cpp
@@ -106,12 +106,13 @@ void sort_pixels_by_bins(double const *const pPixelData,size_t nDataRows,size_t 
     size_t i,j,jBase,ind,i0;
     ppInd[0]=0;
     for(i=1;i<distribution_size;i++){   // calculate the ranges of the cell arrays
-        ppInd[i]=ppInd[i-1]+pCellDens[i-1]; // the next cell starts from the the previous one
+        ppInd[i]=ppInd[i-1]+(size_t)pCellDens[i-1]; // the next cell starts from the the previous one
     };                                      // plus the number of pixels in the cell previous cell
-    if(ppInd[distribution_size-1]+pCellDens[distribution_size-1]!=nDataCols){
+    if(ppInd[distribution_size-1]+(size_t)pCellDens[distribution_size-1]!=nDataCols){
         throw("Sort_pixels_by_bins: pixels data and their cell distributions are inconsistent ");
     }
-    
+
+
 //#pragma omp parallel
 {
     for(j=0;j<nDataCols;j++){    // sort pixels according to cells
